+++ src/main/java/qna/domain/repository/QuestionRepository.java
@@ -1,9 +1,9 @@
-package qna.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package qna.domain.repository;
 
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.Question;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
     List<Question> findByDeletedFalse();
