+++ src/main/java/qna/domain/repository/UserRepository.java
@@ -1,8 +1,8 @@
-package qna.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
+package qna.domain.repository;
 
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import qna.domain.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByUserId(String userId);
